fix: set the button to show based on role

+++ one_compliance/public/js/sales_order.js
@@ -201,34 +201,40 @@ frappe.ui.form.on('Reimbursement Details', {
 });
 
 function handle_unlinking(frm) {
-  if (!frm.is_new()) {
-    frm.add_custom_button("Unlink and Delete", () => {
-      frappe.confirm(
-        __(
-          `Are you sure you want to unlink and delete all linked records of <b>${frm.doc.name}</b>?`
-        ),
-        function () {
-          frappe.call({
-            method:
-              "one_compliance.one_compliance.doc_events.sales_order.delete_linked_records",
-            args: {
-              sales_order: frm.doc.name,
-            },
-            callback: function (r) {
-              if (r.message === "success") {
-                frappe.msgprint(
-                  __("All linked records have been successfully deleted.")
-                );
-                frappe.set_route("List", "Sales Order");
-              } else {
-                frappe.throw(
-                  __("An error occurred while deleting linked records.")
-                );
+  if (!frm.is_new() && frm.doc.workflow_state == "In Progress") {
+    frappe.db
+      .get_single_value("Compliance Settings", "role_allowed_to_unlink_and_delete_sales_orders")
+      .then((role) => {
+        if (role && frappe.user_roles.includes(role)) {
+          frm.add_custom_button("Unlink and Delete", () => {
+            frappe.confirm(
+              __(
+                `Are you sure you want to unlink and delete all linked records of <b>${frm.doc.name}</b>?`
+              ),
+              function () {
+                frappe.call({
+                  method:
+                    "one_compliance.one_compliance.doc_events.sales_order.delete_linked_records",
+                  args: {
+                    sales_order: frm.doc.name,
+                  },
+                  callback: function (r) {
+                    if (r.message === "success") {
+                      frappe.msgprint(
+                        __("All linked records have been successfully deleted.")
+                      );
+                      frappe.set_route("List", "Sales Order");
+                    } else {
+                      frappe.throw(
+                        __("An error occurred while deleting linked records.")
+                      );
+                    }
+                  },
+                });
               }
-            },
+            );
           });
         }
-      );
-    });
+      });
   }
 }
